+++ lib/core/dart_declaration.dart
@@ -5,27 +5,10 @@ import 'package:json_to_model/core/json_model.dart';
 import 'package:json_to_model/utils/extensions.dart';
 import 'package:json_to_model/utils/faker_maker.dart';
 
-abstract class Info<T> {
-  String get title;
-  T get value;
-}
-
-class Replace extends Info<String> {
-  @override
-  final String title = 'DocTypeId';
-  @override
-  final String value;
-
-  Replace(this.value);
-
-  factory Replace.fromJson(dynamic value) {
-    return Replace(value as String);
-  }
-}
-
 class DartDeclaration {
   List<String> imports = [];
   String? type;
+  bool explicitTypeOverride = false;
   dynamic jsonValue;
   String? originalName;
   String? name;
@@ -43,15 +26,6 @@ class DartDeclaration {
   bool get isEnum => enumValues.isNotEmpty;
   bool get isDatetime => type == 'DateTime';
   bool get isTimeStamp => jsonValue == '@timestamp';
-  bool get isInt =>
-      jsonValue is String &&
-      ((jsonValue as String?)?.contains('@int') ?? false);
-  bool get isDouble =>
-      jsonValue is String &&
-      ((jsonValue as String?)?.contains('@double') ?? false);
-  bool get isString => jsonValue is String;
-  bool get isFloat =>  jsonValue is String &&
-      ((jsonValue as String?)?.contains('@float') ?? false);
 
   String get isNullableString => isNullable ? '?' : '';
 
@@ -63,8 +37,7 @@ class DartDeclaration {
   }
 
   String toMockDeclaration(String className) {
-    final value = checkNestedTypes(type!,
-        (String cleanedType, bool isList, bool isListInList, bool isModel) {
+    final value = checkNestedTypes(type!, (String cleanedType, bool isList, bool isListInList, bool isModel) {
       final fakerMaker = FakerMaker(this, className, cleanedType, isModel);
       final fakerDeclaration = fakerMaker.generate();
 
@@ -93,31 +66,25 @@ class DartDeclaration {
       declaration += '@override ';
     }
 
-    declaration +=
-        'final $type$isNullableString $name${stringifyAssignment(assignment)};'
-            .trim();
+    declaration += 'final $type$isNullableString $name${stringifyAssignment(assignment)};'.trim();
 
     return declaration.indented();
   }
 
   String fromJsonBody() {
-    return checkNestedTypes(type!,
-        (String cleanedType, bool isList, bool isListInList, bool isModel) {
+    return checkNestedTypes(type!, (String cleanedType, bool isList, bool isListInList, bool isModel) {
       final jsonVar = "json['$originalName']";
       String conversion;
-      String modelFromJson([String jsonVar = 'e']) =>
-          '$cleanedType.fromJson($jsonVar as Map<String, dynamic>)';
+      String modelFromJson([String jsonVar = 'e']) => '$cleanedType.fromJson($jsonVar as Map<String, dynamic>)';
 
       if (isListInList) {
         conversion =
             '($jsonVar as List? ?? []).map((e) => (e as List? ?? []).map((e) => ${modelFromJson()}).toList()).toList()';
       } else if (isList) {
         if (isModel) {
-          conversion =
-              '($jsonVar as List? ?? []).map((e) => ${modelFromJson()}).toList()';
+          conversion = '($jsonVar as List? ?? []).map((e) => ${modelFromJson()}).toList()';
         } else {
-          conversion =
-              '($jsonVar as List? ?? []).map((e) => e as $cleanedType).toList()';
+          conversion = '($jsonVar as List? ?? []).map((e) => e as $cleanedType).toList()';
         }
       } else if (isModel) {
         conversion = modelFromJson(jsonVar);
@@ -125,18 +92,15 @@ class DartDeclaration {
         conversion = 'DateTime.parse($jsonVar as String)';
       } else if (isTimeStamp) {
         conversion = 'DateTime.fromMillisecondsSinceEpoch($jsonVar as int)';
-      } else if (isString) {
-        cleanedType = (originalName ?? '').toCamelCase();
-        conversion = '$cleanedType.fromJson($jsonVar))';
-      } else if (isInt) {
-        cleanedType = (originalName ?? '').toCamelCase();
-        conversion = '$cleanedType.fromJson($jsonVar))';
-      } else if (isDouble) {
-        cleanedType = (originalName ?? '').toCamelCase();
-        conversion = '$cleanedType.fromJson($jsonVar))';
-      } else if (isFloat) {
-        cleanedType = (originalName ?? '').toCamelCase();
-        conversion = '$cleanedType.fromJson($jsonVar))';
+      } else if (type == 'String') {
+        conversion = '$jsonVar$isNullableString.toString()';
+      } else if (type == 'double') {
+        conversion = '($jsonVar as num).toDouble()';
+      } else if (explicitTypeOverride &&
+          type?.startsWith('Map') != true &&
+          type?.startsWith("List") != true &&
+          type != 'dynamic') {
+        conversion = modelFromJson(jsonVar);
       } else {
         conversion = '$jsonVar as $type';
       }
@@ -150,19 +114,16 @@ class DartDeclaration {
   }
 
   String toJsonBody(String className) {
-    return checkNestedTypes(type!,
-        (String cleanedType, bool isList, bool isListInList, bool isModel) {
+    return checkNestedTypes(type!, (String cleanedType, bool isList, bool isListInList, bool isModel) {
       String conversion;
 
       if (isListInList) {
-        conversion =
-            '$name$isNullableString.map((e) => e.map((e) => e.toJson()).toList()).toList()';
+        conversion = '$name$isNullableString.map((e) => e.map((e) => e.toJson()).toList()).toList()';
       } else if (isList) {
         if (isModel) {
           conversion = '$name$isNullableString.map((e) => e.toJson()).toList()';
         } else {
-          conversion =
-              '$name$isNullableString.map((e) => e.toString()).toList()';
+          conversion = '$name$isNullableString.map((e) => e.toString()).toList()';
         }
       } else if (isModel) {
         conversion = '$name$isNullableString.toJson()';
@@ -195,8 +156,7 @@ class DartDeclaration {
   }
 
   String toCloneDeclaration() {
-    return checkNestedTypes(type!,
-        (String cleanedType, bool isList, bool isListInList, bool isModel) {
+    return checkNestedTypes(type!, (String cleanedType, bool isList, bool isListInList, bool isModel) {
       if (isListInList) {
         return '$name: $name${isNullable ? '?' : ''}.map((x) => x.map((y) => y.clone()).toList()).toList()';
       } else if (isList) {
@@ -228,14 +188,9 @@ class DartDeclaration {
       }
     }
 
-    final importExists = imports.indexWhere(
-            (element) => element.contains(cleanType.toSnakeCase())) !=
-        -1;
-    final nestedClassExists =
-        nestedClasses.indexWhere((element) => element.className == cleanType) !=
-            -1;
-    final isModel =
-        !isEnum && (importExists || nestedClassExists || this.isModel);
+    final importExists = imports.indexWhere((element) => element.contains(cleanType.toSnakeCase())) != -1;
+    final nestedClassExists = nestedClasses.indexWhere((element) => element.className == cleanType) != -1;
+    final isModel = !isEnum && (importExists || nestedClassExists || this.isModel);
 
     return callback(cleanType, isList, isListInList, isModel);
   }
@@ -303,7 +258,7 @@ class DartDeclaration {
     if (import == null) {
       return;
     }
-    if (import is List && !import.isNotEmpty) {
+    if (import is List && import.isNotEmpty) {
       imports.addAll(import.map((e) => e as String));
     } else if (import != null && import is String) {
       imports.add(import);
@@ -356,17 +311,14 @@ class DartDeclaration {
           final commandPrefixMatch = command.prefix != null &&
               command.command != null &&
               testSubject.startsWith(command.prefix! + command.command!);
-          final commandMatch = command.command == null ||
-              testSubject.startsWith(command.command!);
+          final commandMatch = command.command == null || testSubject.startsWith(command.command!);
 
           if (commandPrefixMatch || commandMatch) {
             final notprefixnull = command.notprefix == null;
-            final notprefix =
-                !notprefixnull && !testSubject.startsWith(command.notprefix!);
+            final notprefix = !notprefixnull && !testSubject.startsWith(command.notprefix!);
 
             if (notprefix || notprefixnull) {
-              newSelf =
-                  command.callback(self, testSubject, key: key, value: value);
+              newSelf = command.callback(self, testSubject, key: key, value: value);
               break;
             }
           }
@@ -472,5 +424,4 @@ String _detectType(String value) {
   return 'String';
 }
 
-typedef NestedCallbackFunction = String Function(
-    String cleanedType, bool isList, bool isListInList, bool isModel);
+typedef NestedCallbackFunction = String Function(String cleanedType, bool isList, bool isListInList, bool isModel);